+++ backend/src/main/kotlin/uk/co/threebugs/darwinexclient/trade/TradeDto.kt
@@ -1,10 +1,10 @@
 package uk.co.threebugs.darwinexclient.trade
 
-import uk.co.threebugs.darwinexclient.Status
-import uk.co.threebugs.darwinexclient.account.AccountDto
-import uk.co.threebugs.darwinexclient.setup.SetupDto
-import java.math.BigDecimal
-import java.time.ZonedDateTime
+import uk.co.threebugs.darwinexclient.*
+import uk.co.threebugs.darwinexclient.account.*
+import uk.co.threebugs.darwinexclient.setup.*
+import java.math.*
+import java.time.*
 
 
 data class TradeDto(
@@ -14,7 +14,7 @@ data class TradeDto(
     var status: Status,
     val setup: SetupDto,
     val account: AccountDto,
-    var metatraderId: Int? = null,
+    var metatraderId: Long? = null,
     var targetPlaceDateTime: ZonedDateTime? = null,
     var placedDateTime: ZonedDateTime? = null,
     var placedPrice: BigDecimal? = null,
